+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Compatible with `jj` 0.37.0.
 * Updated colors to make output easier to read.
 * Updated field order to match `jj` revset engine internals.
 * Renamed operations to better match `jj` revset engine internals.
+* Renamed `--no-config` to `--no-user-config`.
 
 ## [0.1.0] - 2026-01-12